Fix for allowing canceling of signin.
AuthenticateActivity was starting the browser
in onCreate then calling finish. This doesn't
allow onSaveInstanceState to be called. If this
activity is killed then state is lost. When
the browser returns to this activity it will
start the browser again since it doesnt know
its previous state.. Fix is to start browser
in onResume allowing onSaveInstanceState to
be called if the activity is killed.