Storage is blocking. Make sure data is stored.
When launching the brwoser the activity might
be killed. This causes a bug where the state
is lost because the background thread is not
finished. To fix we use commit instead of
apply to make sure data is saved.
An async method can be implemented
in OktaRepository if saving data in a
background thread is needed.