Fix bug when switching between encryption managers
Bug was caused by removing keys and reusing the same
encryption manager object. The keys need to be recreated.

Add sample to show how to combine the encryption manager
with the keyguard. This will require the user to authenticate
the device before any profile or token calls could be done.
Espresso test added.