HttpResponse: Throw HttpStatusCodeException when unexpected status code is encountered.
Prerequisite to: OKTA-282915

Adds a new HttpStatusCodeException that subclasses IOException that is thrown when an unexpected status code is encountered on parsing the response as JSON, in place of a plain IOException.

This change is transparent to existing consumers that are unaware of HttpStatusCodeException. The underlying exception message remains the same as before.